Get live camera view working
